+++ components/FormWizard/DynamicStep.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import { useForm } from 'react-hook-form';
 
 import Button from 'components/Button/Button';
-import DynamicInput from 'components/DynamicStep/DynamicInput';
+import DynamicInput from 'components/FormWizard/DynamicInput';
 
 const DynamicStep = ({
   isMulti,
